+++ lib/td_df_lib/map_diff.ex
@@ -7,6 +7,9 @@ defmodule TdDfLib.MapDiff do
   Returns a key-wise difference between two maps, considering the value `nil` to
   be equivalent to an empty map.
 
+  A mask may be applied to the diff values by specifying the `mask` option (a
+  1-arity function).
+
   ## Examples
 
       iex> MapDiff.diff(%{foo: 1, bar: 2, baz: 3}, %{foo: 1, bar: 22, xyzzy: 3})
@@ -17,30 +20,38 @@ defmodule TdDfLib.MapDiff do
       }
 
   """
-  def diff(old, new)
+  def diff(old, new, opts \\ [])
 
-  def diff(nil, nil), do: %{}
-  def diff(nil, %{} = content), do: diff(%{}, content)
-  def diff(%{} = content, nil), do: diff(content, %{})
+  def diff(nil, nil, _opts), do: %{}
+  def diff(nil, %{} = content, opts), do: diff(%{}, content, opts)
+  def diff(%{} = content, nil, opts), do: diff(content, %{}, opts)
 
-  def diff(%{} = old, %{} = new) do
+  def diff(%{} = old, %{} = new, opts) do
     %{
-      added: key_diff(old, new),
-      removed: key_diff(new, old),
-      changed: val_diff(old, new)
+      added: key_diff(old, new, opts[:mask]),
+      removed: key_diff(new, old, opts[:mask]),
+      changed: val_diff(old, new, opts[:mask])
     }
     |> Enum.reject(fn {_, map} -> map == %{} end)
     |> Map.new()
   end
 
-  defp key_diff(%{} = old, %{} = new) do
-    Map.drop(new, Map.keys(old))
+  defp key_diff(%{} = old, %{} = new, mask_fn) do
+    new
+    |> Map.drop(Map.keys(old))
+    |> to_map(mask_fn)
   end
 
-  defp val_diff(%{} = old, %{} = new) do
+  defp val_diff(%{} = old, %{} = new, mask_fn) do
     new
     |> Map.take(Map.keys(old))
     |> Enum.reject(fn {k, v} -> Map.get(old, k) == v end)
-    |> Map.new()
+    |> to_map(mask_fn)
+  end
+
+  defp to_map(enumerable, nil), do: Map.new(enumerable)
+
+  defp to_map(enumerable, f) when is_function(f, 1) do
+    Map.new(enumerable, fn {k, v} -> {k, f.(v)} end)
   end
 end